Set IDbDataParameter.DbType when the value is null.
I'm actually trying to track down an InvalidOperationException from EntitySet;
see the new WriteTest.InsertAndDeleteWithDependencies() test for the example.

However, InsertAndDeleteWithDependencies() was failing _before_ it hit the 
EntitySet bug, as it was trying to insert a null string for Category.Picture
instead of a null Binary type.  This caused no end of fail.

	* src/DbLinq/Test/Providers/WriteTest.cs: Add
	  InsertAndDeleteWithDependencies() test, to elicit
	  InvalidOperationException from EntitySet bug.  This passes under
	  DbLinq.SqlServer_test_strict.dll, but fails for DbLinq.
	* src/DbLinq/Util/IDbDataParameterExtensions.cs: If the value is null,
	  attempt to set IDbDataParameter.DbType based on the actual type.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1069 33146b6e-df29-0410-a1ca-39be37fb9416