SqlProvider.GetLitearl(DateTime) shouldn't depend on the user's locale.
User locale is bad ju-ju.  (M/D/Y or D/M/Y?  Who knows?  Especially when the
DB is on a different box from the DbLinq machine (e.g. ASP.NET sites), they
may not even be in the same time zone.  Far better/saner as the default to
return the "o" format, which is YYYY-MM-DDTHH:MM:SS.SSS, which is an ISO
standard and won't be confused with anything else.)


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1109 33146b6e-df29-0410-a1ca-39be37fb9416