"Storage" property implementation
When "Storage" is specified for a Column or Association attribute the data is
directly stored into the backing field instead of going through the setter
method of the annotated property.

Also added a test that uses a custom version of Northwind tables to test for
the behaviour of the code with and without the "Storage" attribute and with
or without a setter method for the property.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1185 33146b6e-df29-0410-a1ca-39be37fb9416