Fix failing DbLinq.SqlServer_test.dll regressions: ReadTest.Storage01() and ReadTest_Complex.F27_SelectEmployee_Identifier().
The problem was that the e.g. NoStorageCategory.CategoryID property had no
setter, so PropertyInfo.GetSetMethod() was returning null, resulting in a
ArgumentNullException from Delegate.CreateDelegate().

The fix is to use the ColumnAttribute.Storage value if the setter is null, and
fallback to setting the field value instead of using the property setter.

(We also make these types internal, as they shouldn't be exported from
System.Data.Linq.dll.)


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1212 33146b6e-df29-0410-a1ca-39be37fb9416