Test DbSchemaLoader behavior.
Caveat: it only works for SQLite, because in order to sanely work we need
foreign-key information, and only SQLite actually inserts FK information into
its DataTable output.  Still, it serves as a useful base for many things.

Aside: this looks gigantic because it does three related things:

 1. It adds a test to CreateEntitiesFromSqliteDbTest.cs to check the
    DbSchemaLoader behavior.  This also adds "expected output" to
    tests/expected.
 2. It adds DbSchemaLoader to System.Data.Linq.dll (so that the
    SqlMetal_test_sqlite.dll test works).
 3. It changes the SQLite Northwind definition.

(3) is the biggest change, and is due to this: Northwind isn't consistent.
Specifically, it had:

    CREATE TABLE Employees (
        EmplyeeID INTEGER NOT NULL, ; PRIMARY KEY
        ; ...
    );
    CREATE TABLE Orders (
        EmployeeID "int" NULL,      ; FOREIGN KEY to Employees
        ; ...
    );

This worked because all SQL type -> managed type mapping was done by
SchemaLoader.MapDbType(), so both INTEGER and int mapped to System.Int32, and
everything was kosher.

HOWEVER, DbSchemaLoader doesn't use SchemaLoader.MapDbType() (because it's
evil), and instead uses the DbConnection's DataTypes DataTable to perform the
SQL type -> managed type mapping.

This sounds rational.

Problem was, System.Data.SQLite.dll was mapping INTEGER to System.Int64, not
System.Int32 (long vs. int).

Result: the generated entities wouldn't compile, because you can't implicitly
assign a long to an int. :-)

Obviously, this is because the database schema is inconsistent -- if a column
needs to store an INTEGER, it better itself be an INTEGER.  So we go off and
fix the Northwind SQLite schema, which introduces changes in the expected
output for normal SQLite schema generation, etc.

It also necessitated one change in ReadTest_Complex.cs because the test was
assuming that Product.ProductID was an int, but it's actually a long under
DbSchemaLoader output, so it wouldn't compile.  The cast fixes that.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1284 33146b6e-df29-0410-a1ca-39be37fb9416