DbMetal: Allow intermixing of /provider and --with-* options.
Previously, if you wanted to use a custom --with-schema-loader, you had to
provide ALL of --with-schema-loader=T, --with-db-connection=T, and
--with-sql-dialect, even if --provider could provide --with-db-connection (and
if we extend the .exe.config file, it could provide --with-sql-dialect too).

Instead, treat the --with-* options as overrides -- if --provider is present,
load the specified provider, then override the default .exe.config values with
the --with-* values.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1303 33146b6e-df29-0410-a1ca-39be37fb9416