Fix storage and member aliasing.
(Mutable global data is bad, mkay?)

Because theoriginal code was updating `association.Storage` and
`association.Member` (the "global data" referred to above), this was
*reducing* conflicts.  Unfortunately, this is bad because it caused other
sections of the code to not use the *original* values (as the original values
were lost), thus resulting in broken aliases.

Patch thanks to andershol.

Fixes issue 196.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1309 33146b6e-df29-0410-a1ca-39be37fb9416