Refactor the DbMetal/SqlMetal tests.
I want to add more tests for DbMetal, the logic for which is mostly identical
to the existing tests (it just changes the command-line arguments and the
expected filename).  Instead of rampant copy/pasting, cleanup the
DbMetal/SqlMetal tests so that logic can be more easily used.

As part of this process, "regularize" the placement of "expected output" files
so that they're placed within tests/expected, instead of some being
tests/*.Expected*, and others being in tests/expected/*.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1325 33146b6e-df29-0410-a1ca-39be37fb9416