Use a custom XML Serializer.
Use of XmlSerializer to generate DBML causes 'sqlmetal' to fail, as
XmlSerializer requires that the types be public, which they can't be.

Instead, use an 'sgen'-generated XmlSerializer implementation, which thus
avoids the need for public types, and allows sqlmetal to work.

Add tests for DBML generation from DbMetal and SqlMetal.

DbmlSerializer.cs file thanks to jiaguilera.

Fixes Issue 171.


git-svn-id: http://dblinq2007.googlecode.com/svn/trunk@1326 33146b6e-df29-0410-a1ca-39be37fb9416